+++ ScopeControl/Changelog.md
@@ -1,5 +1,9 @@
 ﻿# Changelog
 
+## Version 1.4.0.0
+- Fixed moon button toggle not being persistent when coordinates change
+- Added an azimuth chart
+
 ## Version 1.2.3.0
 - Added a start/stop tracking button
 - Rearranged the manual coordinate slew controls to switch between ra/dec and alt/az instead of showing both together